securex

+++ components/canodrome/controller.js
@@ -2,6 +2,7 @@ const store = require('./store');
 const storeUser = require('../user/store');
 const storeCan = require('../cans/store');
 const { mint } = require('../../services/nftCanodromes.js');
+const req = require('express/lib/request');
 const socket = require('../../socket').socket;
 
 //GET CANODROMES USER
@@ -148,6 +149,12 @@ const sellCanodrome = (canodromeId, canodrome) => {
         try {
             
             canodrome.cans = [];
+
+            //check
+            const canodromeCheck = await store.get(canodromeId);
+
+            if(canodromeCheck.wallet != canodrome.wallet.toLowerCase()) throw 'Not authorized';
+
             const newCanodrome = await store.setSellCanodrome(canodromeId, canodrome);
             
             //websocket
@@ -166,6 +173,11 @@ const sellCanodrome = (canodromeId, canodrome) => {
 const removeCanodrome = canodromeId => {
     return new Promise( async (resolve, reject) => {
         try {
+
+            //check
+            const canodromeCheck = await store.get(canodromeId);
+
+            if(canodromeCheck.wallet != req.user.wallet.toLowerCase()) throw 'Not authorized';
 
             const newCanodrome = await store.setRemoveCanodrome(canodromeId);
 
@@ -186,6 +198,11 @@ const setStatusCanodrome = (canodromeId, status) => {
     return new Promise( async (resolve, reject) => {
         try {
 
+            //check
+            const canodromeCheck = await store.get(canodromeId);
+
+            if(canodromeCheck.wallet != req.user.wallet.toLowerCase()) throw 'Not authorized';
+
             const newCanodrome = await store.setStatus(canodromeId, status);
 
             resolve(newCanodrome);
@@ -202,7 +219,11 @@ const deleteCan = (canodromeId, canId) => {
     return new Promise( async (resolve, reject) => {
         try {
           
-            const canodrome = await get(canodromeId); 
+            //check
+            const canodrome = await get(canodromeId);
+
+            if(canodrome.wallet != req.user.wallet.toLowerCase()) throw 'Not authorized';
+            
             const arrayCans = canodrome.cans.filter(can => can.can.id != canId)      
             const exe = await store.delete(canodromeId, arrayCans);
             resolve(exe);
